end--no-edit

+++ Jobsheet2/mahasiswaMain23.java
@@ -4,7 +4,7 @@ public class mahasiswaMain23 {
         mhs1.nama = "Muhammad Ali Farhan";
         mhs1.nim = "2241720171";
         mhs1.kelas = " SI 2J";
-        mhs1.ipk = 3.55;
+        mhs1.ipk = (float)3.55;
 
         mhs1.tampilkanInformasi();
         mhs1.ubahKelas("SI 2K");
